pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "discolike-cli"
version = "0.1.0"
description = "CLI for the DiscoLike B2B company discovery API"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "httpx>=0.27",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "respx>=0.21",
    "ruff>=0.4",
    "mypy>=1.10",
]

[project.scripts]
discolike = "discolike.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/discolike"]

[tool.ruff]
src = ["src", "tests"]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[[tool.mypy.overrides]]
module = "respx.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "yaml.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "discolike.cli"
# Click decorators produce dynamic types that mypy can't resolve
disable_error_code = ["has-type"]

[[tool.mypy.overrides]]
module = "discolike.commands.*"
# Click's pass_context decorator has known mypy compat issues
disable_error_code = ["arg-type"]

[tool.pytest.ini_options]
testpaths = ["tests"]
